Remove rust from the supported languages
After giving it some thought, it seems that Rust would be more trouble than it is currently worth for supporting at this time. The compile times increased 6x total with no code added besides basic kernel bindings. Additionally, where memory is allocated inside of the kernel, it normally is never freed because it is eternal, or is allocated in D, where the destructor deallocates memory.