add writev(2): actual buffered writing for stdio!
This commit adds the writev syscall, which alows multiple entire buffers to be written to a file at a time. This is used in stdio for buffered writing to the console. We flush everytime printf is used and/or after 512 invocations of something writing to the console, such as putc or putchar.