Add signals support
with these changes, there is now basic signals support. One small difference might be that SIGFPE cannot be handled. It is also hardcoded that you cannot kill the init or first shell process with ctrl-c, but you can kill any other program with it.

There is also some ksyms work in Rust. It isn't harmful to commit it, but it currently does nothing. Eventually, it will be used to parse ksyms.map for better panic messages.