Move code from sysfile.c into file.c
this helps to provide a better separation of purposes in the files. sysfile.c is for implementations of syscalls having to do with files, and file.c is for the implementation of the core functions that those syscalls use. sys_open uses fileopen and sys_close uses fileclose for example. This is the closest we have to a VFS at the moment.